fuzz delegation - rewards, withdraw, check balances, refactoring
